academia/pyproject.toml
[project]
name = "tecnologia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "psycopg[binary,binary-pool,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
]
